+++ include/rcl_cloud_action.h
@@ -226,7 +226,7 @@ class RCloudAction
                 static const QString key;
                 static const QString description;
             };
-            struct Query
+            struct AIQuery
             {
                 static const QString key;
                 static const QString description;