+++ src/rcl_cloud_action.cpp
@@ -122,8 +122,8 @@ const QString RCloudAction::Action::ProcessUpdateAccessMode::description = "Upda
 const QString RCloudAction::Action::SubmitReport::key = "submit-report";
 const QString RCloudAction::Action::SubmitReport::description = "Submit report to the cloud server";
 
-const QString RCloudAction::Action::Query::key = "query";
-const QString RCloudAction::Action::Query::description = "Send a query to the cloud server";
+const QString RCloudAction::Action::AIQuery::key = "ai-query";
+const QString RCloudAction::Action::AIQuery::description = "Submit an AI query to be answered by an external AI service";
 
 void RCloudAction::_init(const RCloudAction *pCAction)
 {
@@ -269,7 +269,7 @@ QMap<QString, QString> RCloudAction::getActionMap()
     actionMap.insert(Action::ProcessUpdateAccessOwner::key,Action::ProcessUpdateAccessOwner::description);
     actionMap.insert(Action::ProcessUpdateAccessMode::key,Action::ProcessUpdateAccessMode::description);
     actionMap.insert(Action::SubmitReport::key,Action::SubmitReport::description);
-    actionMap.insert(Action::Query::key,Action::Query::description);
+    actionMap.insert(Action::AIQuery::key,Action::AIQuery::description);
 
     return actionMap;
 }